requirements.txt
discord.py==1.2.5
loguru
pip
GitPython
sqlalchemy
alembic
mysqlclient
ConfigParser
